Feature/86cw24p8f Util function to get logged-in user info (#78)
feature/86cw24p8f: Created util function to get logged in user information easily
