+++ behind-the-veil-siteroot/imports/ui/components/pages/profile/ProfilePage.jsx
@@ -4,18 +4,18 @@
  * Contributors: Nikki
  */
 
-import React, {useState} from "react";
+import React from "react";
 import {useNavigate} from "react-router-dom";
-import {Tracker} from "meteor/tracker";
-import {Meteor} from "meteor/meteor";
+import {Cog8ToothIcon} from "@heroicons/react/24/outline";
+
 import WhiteBackground from "../../whiteBackground/WhiteBackground";
 import PageLayout from "../../../enums/PageLayout";
 import Button from "../../button/Button";
-import {Cog8ToothIcon} from "@heroicons/react/24/outline";
 import ProfileDisplay from "../../profilePhoto/ProfileDisplay";
 import BrideProfileTabs from "./BrideProfileTabs";
 import ArtistProfileTabs from "./ArtistProfileTabs";
 import UrlBasePath from "../../../enums/UrlBasePath";
+import {getUserInfo} from "../../util";
 
 /**
  * The general profile page (changes for content depending on user type)
@@ -25,32 +25,7 @@ export const ProfilePage = () => {
     const navigate = useNavigate();
 
     // get current user information
-    const [userInfo, setUserInfo] = useState(
-        {"alias": null, "username": null, "type": null}
-    );
-
-    // tracker for the required user data updates
-    Tracker.autorun(() => {
-        const user = Meteor.user();
-
-        if (user) {
-            // user data is returned (sometimes it takes a while)
-            const username = user.username;
-            const userAlias = user.profile.alias;
-            const userType = user.profile.type;
-
-            // check if an update to the current user info is required or not (this is needed to prevent inf loop)
-            if (userInfo.username !== username || userInfo.alias !== userAlias || userInfo.type !== userType) {
-                setUserInfo(
-                    {
-                        "alias": user.profile.alias,
-                        "type": user.profile.type,
-                        "username": user.username
-                    }
-                )
-            }
-        }
-    })
+    const userInfo = getUserInfo();
 
     return (
         <WhiteBackground pageLayout={PageLayout.LARGE_CENTER}>
